V0.11.0 (#157)
* adjustments for implicite solvers

* major refactoring
