+++ Project.toml
@@ -1,7 +1,7 @@
 name = "FMI"
 uuid = "14a09403-18e3-468f-ad8a-74f8dda2d9ac"
 authors = ["TT <tobias.thummerer@informatik.uni-augsburg.de>", "LM <lars.mikelsons@informatik.uni-augsburg.de>", "JK <josef.kircher@student.uni-augsburg.de>"]
-version = "0.10.2"
+version = "0.11.0"
 
 [deps]
 ChainRulesCore = "d360d2e6-b24c-11e9-a2a3-2a2ae2dbcce4"
@@ -20,7 +20,7 @@ ChainRulesCore = "1.15.0"
 DiffEqCallbacks = "2.24.0"
 DifferentialEquations = "7.5.0"
 FMIExport = "0.1.0"
-FMIImport = "0.12.1"
+FMIImport = "0.13.0"
 ForwardDiff = "0.10.0"
 ProgressMeter = "1.7.0"
 Requires = "1.3.0"
