+++ test/FMI2/cs_me.jl
@@ -21,9 +21,9 @@ elseif envFMUSTRUCT == "FMUCOMPONENT"
 end
 @assert fmuStruct !== nothing "Unknown fmuStruct, environment variable `FMUSTRUCT` = `$envFMUSTRUCT`"
 
-sol = fmiSimulateCS(fmuStruct, t_start, t_stop)
+sol = fmiSimulateCS(fmuStruct, (t_start, t_stop))
 @test sol.success 
-sol = fmiSimulateME(fmuStruct, t_start, t_stop)
+sol = fmiSimulateME(fmuStruct, (t_start, t_stop))
 @test sol.success 
 fmiUnload(myFMU)
 
@@ -39,7 +39,7 @@ if envFMUSTRUCT == "FMU"
 elseif envFMUSTRUCT == "FMUCOMPONENT"
     fmuStruct = comp
 end
-sol = fmiSimulate(fmuStruct, t_start, t_stop)
+sol = fmiSimulate(fmuStruct, (t_start, t_stop))
 @test sol.success
 fmiUnload(myFMU)
 
@@ -55,6 +55,6 @@ if envFMUSTRUCT == "FMU"
 elseif envFMUSTRUCT == "FMUCOMPONENT"
     fmuStruct = comp
 end
-sol = fmiSimulate(fmuStruct, t_start, t_stop)
+sol = fmiSimulate(fmuStruct, (t_start, t_stop))
 @test sol.success
 fmiUnload(myFMU)
