+++ test/FMI2/load_save.jl
@@ -13,8 +13,8 @@ t_stop = 8.0
 myFMU = fmiLoad("SpringFrictionPendulum1D", ENV["EXPORTINGTOOL"], ENV["EXPORTINGVERSION"])
 
 recordValues = ["mass.s", "mass.v"]
-solutionME = fmiSimulateME(myFMU, t_start, t_stop; recordValues=recordValues)
-solutionCS = fmiSimulateCS(myFMU, t_start, t_stop; recordValues=recordValues)
+solutionME = fmiSimulateME(myFMU, (t_start, t_stop); recordValues=recordValues)
+solutionCS = fmiSimulateCS(myFMU, (t_start, t_stop); recordValues=recordValues)
 
 # ME
