+++ test/FMI2/plots.jl
@@ -19,8 +19,8 @@ fmiInfo(myFMU)
 fmiInstantiate!(myFMU)
 
 recordValues = ["mass.s", "mass.v"]
-solutionME = fmiSimulateME(myFMU, t_start, t_stop; recordValues=recordValues)
-solutionCS = fmiSimulateCS(myFMU, t_start, t_stop; recordValues=recordValues)
+solutionME = fmiSimulateME(myFMU, (t_start, t_stop); recordValues=recordValues)
+solutionCS = fmiSimulateCS(myFMU, (t_start, t_stop); recordValues=recordValues)
 
 # plot the results
 fig = fmiPlot(solutionME)
