+++ test/FMI2/sim_CS.jl
@@ -24,11 +24,11 @@ t_start = 0.0
 t_stop = 8.0
 
 # test without recording values (but why?)
-solution = fmiSimulateCS(fmuStruct, t_start, t_stop; dt=1e-2)
+solution = fmiSimulateCS(fmuStruct, (t_start, t_stop); dt=1e-2)
 @test solution.success
 
 # test with recording values
-solution = fmiSimulateCS(fmuStruct, t_start, t_stop; dt=1e-2, recordValues=["mass.s", "mass.v"])
+solution = fmiSimulateCS(fmuStruct, (t_start, t_stop); dt=1e-2, recordValues=["mass.s", "mass.v"])
 @test solution.success
 @test length(solution.values.saveval) == t_start:1e-2:t_stop |> length
 @test length(solution.values.saveval[1]) == 2
@@ -82,7 +82,7 @@ end
 for inpfct in [extForce_cxt, extForce_t]
     global solution
 
-    solution = fmiSimulateCS(fmuStruct, t_start, t_stop; dt=1e-2, recordValues=["mass.s", "mass.v"], inputValueReferences=["extForce"], inputFunction=extForce_t)
+    solution = fmiSimulateCS(fmuStruct, (t_start, t_stop); dt=1e-2, recordValues=["mass.s", "mass.v"], inputValueReferences=["extForce"], inputFunction=extForce_t)
     @test solution.success
     @test length(solution.values.saveval) > 0
     @test length(solution.values.t) > 0
