+++ test/runtests.jl
@@ -26,65 +26,50 @@ function runtests(exportingTool)
 
     @testset "Testing FMUs exported from $exportingTool" begin
 
-        @info "Sensitivities (sens.jl)"
-        @testset "Sensitivities" begin
-            include("sens.jl")
-        end
-
-        @info "Model Description (model_description.jl)"
-        @testset "Model Description" begin
-            include("model_description.jl")
-        end
-
         for str in fmuStructs
             @testset "Functions for $str" begin
                 ENV["FMUSTRUCT"] = str
 
                 @info "Variable Getters / Setters (getter_setter.jl)"
                 @testset "Variable Getters / Setters" begin
-                    include("getter_setter.jl")
+                    include("FMI2/getter_setter.jl")
                 end
 
                 @info "State Manipulation (state.jl)"
                 @testset "State Manipulation" begin
-                    include("state.jl")
-                end
-
-                @info "Directional derivatives (dir_ders.jl)"
-                @testset "Directional derivatives" begin
-                    include("dir_ders.jl")
+                    include("FMI2/state.jl")
                 end
 
                 @info "Automatic Simulation (sim_auto.jl)"
                 @testset "Automatic Simulation (CS or ME)" begin
-                    include("sim_auto.jl")
+                    include("FMI2/sim_auto.jl")
                 end
 
                 @info "CS Simulation (sim_CS.jl)"
                 @testset "CS Simulation" begin
-                    include("sim_CS.jl")
+                    include("FMI2/sim_CS.jl")
                 end
 
                 @info "ME Simulation (sim_ME.jl)"
                 @testset "ME Simulation" begin
-                    include("sim_ME.jl")
+                    include("FMI2/sim_ME.jl")
                 end
 
                 @info "Support CS and ME simultaneously (cs_me.jl)"
                 @testset "Support CS and ME simultaneously" begin
-                    include("cs_me.jl")
+                    include("FMI2/cs_me.jl")
                 end
 
                 @info "Loading/Saving simulation results (load_save.jl)"
                 @testset "Loading/Saving simulation results" begin
-                    include("load_save.jl")
+                    include("FMI2/load_save.jl")
                 end
             end
         end
 
         @info "Plotting (plots.jl)"
         @testset "Plotting" begin
-            include("plots.jl")
+            include("FMI2/plots.jl")
         end
     end
 end
